Refactor modules.
